requirements.txt
asgiref==3.8.1
blinker==1.6.3
Brotli==1.1.0
cffi==1.17.1
click==8.1.7
cssselect2==0.7.0
distlib==0.3.8
Django==5.1.1
django-cors-headers==4.6.0
env==0.1.0
filelock==3.15.4
Flask==2.3.3
fonttools==4.55.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==9.5.0
pillow==10.4.0
platformdirs==4.2.2
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyphen==0.15.0
sqlparse==0.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
virtualenv==20.26.3
weasyprint==63.0
webencodings==0.5.1
Werkzeug==2.3.7
zopfli==0.2.3.post1
